DFE-1175 - splitting create and edit release services prior to wiring up Create Release to real APIs
