+++ src/explore-education-statistics-admin/src/services/release/create-release/types.ts
@@ -1,0 +1,5 @@
+import {BaseReleaseSetupDetailsRequest} from "@admin/services/release/types";
+
+export interface CreateReleaseRequest extends BaseReleaseSetupDetailsRequest {
+  publicationId: string;
+}
