+++ src/explore-education-statistics-admin/src/services/release/edit-release/setup/mock/mock-data.ts
@@ -1,4 +1,4 @@
-import { ReleaseSetupDetails } from '@admin/services/edit-release/setup/types';
+import {ReleaseSetupDetails} from "@admin/services/release/types";
 import { Dictionary } from '@common/types';
 
 const setupByReleaseId: Dictionary<ReleaseSetupDetails> = {
