+++ src/explore-education-statistics-admin/src/services/release/edit-release/setup/service.ts
@@ -1,8 +1,5 @@
-import {
-  CreateReleaseRequest,
-  ReleaseSetupDetails,
-  UpdateReleaseSetupDetailsRequest,
-} from '@admin/services/edit-release/setup/types';
+import {ReleaseSetupDetails} from '@admin/services/release//types';
+import {UpdateReleaseSetupDetailsRequest} from "@admin/services/release/edit-release/setup/types";
 import { createClient } from '@admin/services/util/service';
 import mocks from './mock/mock-service';
 
@@ -11,9 +8,6 @@ const apiClient = createClient({
 });
 
 export interface ReleaseSetupService {
-  createRelease: (
-    createRequest: CreateReleaseRequest,
-  ) => Promise<ReleaseSetupDetails>;
   getReleaseSetupDetails: (releaseId: string) => Promise<ReleaseSetupDetails>;
   updateReleaseSetupDetails: (
     updatedRelease: UpdateReleaseSetupDetailsRequest,
@@ -21,16 +15,6 @@ export interface ReleaseSetupService {
 }
 
 const service: ReleaseSetupService = {
-  createRelease(
-    createRequest: CreateReleaseRequest,
-  ): Promise<ReleaseSetupDetails> {
-    return apiClient.then(client =>
-      client.post(
-        `/publication/${createRequest.publicationId}/releases`,
-        createRequest,
-      ),
-    );
-  },
   getReleaseSetupDetails(releaseId: string): Promise<ReleaseSetupDetails> {
     return apiClient.then(client =>
       client.get<ReleaseSetupDetails>(`/release/${releaseId}/setup`),
