+++ src/explore-education-statistics-admin/src/services/release/edit-release/setup/types.ts
@@ -1,0 +1,6 @@
+import {BaseReleaseSetupDetailsRequest} from "@admin/services/release/types";
+
+export interface UpdateReleaseSetupDetailsRequest
+  extends BaseReleaseSetupDetailsRequest {
+  releaseId: string;
+}
